Move com.synnapps.carouselview into repository
The project is no longer maintained (last update 4 years ago).
Unfortunately, it was hosted on the JCenter repository, which has been
deprecated. We need to switch away from all use of JCenter in order to
upgrade to Gradle 8. Because this library is not maintained, it hasn't
been migrated to a new repository, so this copies the source into this
project as a module so that we can continue to use it (and update it if
it breaks).

It's a good thing that Gradle exposed this issue because the JCenter
repository is slated to be completely shut down eventually:
https://blog.gradle.org/jcenter-shutdown#impact-builds This would have
prevented this project from being built if we hadn't migrated away from
JCenter by the time it happened. It supposedly was shut down in February
2022, but it must be still available in some sort of read-only mode
given that we've spun up this project from scratch relatively recently.